requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
influxdb-client==1.43.0
paho-mqtt==2.1.0
python-dotenv==1.0.1
requests==2.32.3
